Move util.map to util.ui
