+++ app/src/main/java/com/groupseven/musicmap/screens/main/artist/ArtistDataMapFragment.java
@@ -9,7 +9,7 @@ import com.groupseven.musicmap.models.Artist;
 import com.groupseven.musicmap.models.User;
 import com.groupseven.musicmap.screens.main.map.MapFragment;
 import com.groupseven.musicmap.util.firebase.Queries;
-import com.groupseven.musicmap.util.map.MusicMemoryOverlay;
+import com.groupseven.musicmap.util.ui.map.MusicMemoryOverlay;
 
 import org.osmdroid.views.overlay.FolderOverlay;
 import org.osmdroid.views.overlay.Overlay;
