+++ app/src/main/java/com/groupseven/musicmap/util/ui/map/MusicMemoryOverlay.java
@@ -1,4 +1,4 @@
-package com.groupseven.musicmap.util.map;
+package com.groupseven.musicmap.util.ui.map;
 
 import android.content.Context;
 import android.content.Intent;
